backend/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
colorama==0.4.6
cors==1.0.1
filelock==3.12.2
Flask==2.3.3
Flask-Cors==4.0.0
fsspec==2023.6.0
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==8.0.0
PySocks==1.7.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
safetensors==0.3.3
semantic-version==2.10.0
setuptools-rust==1.7.0
six==1.16.0
sympy==1.12
tldextract==3.4.4
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
